Extract front-end into keel-syntax crate
Move the lexer, AST, parser, formatter, and linter into a new
crates/keel-syntax workspace crate and re-export them from keel-lang
under their original module paths, so the existing crate:: imports
across the codebase compile unchanged.

This is the first step of the workspace split (#71). The single
keel-lang crate was the heaviest compile unit in the whole graph and
was rebuilt on every source change. Isolating the front end cuts the
parser-edit test build from 6.9s to 2.6s, lets front-end-only test
builds finish in under a second, and stops interpreter edits from
rebuilding the front end.

LintWarning moves into keel-syntax alongside the linter; the public
diagnostics facade re-exports it. The formatter idempotency test's
project_root() now ascends to the repo root to locate examples/.
