+++ Cargo.toml
@@ -1,5 +1,10 @@
 [workspace]
-members = ["tools/mdbook-keel-catalog"]
+members = ["crates/*", "tools/mdbook-keel-catalog"]
+
+[workspace.dependencies]
+logos = "0.14"
+chumsky = "0.9"
+miette = { version = "7", features = ["fancy"] }
 
 [package]
 name = "keel-lang"
@@ -19,10 +24,11 @@ name = "keel"
 path = "src/main.rs"
 
 [dependencies]
-logos = "0.14"
-chumsky = "0.9"
+keel-syntax = { path = "crates/keel-syntax" }
+logos = { workspace = true }
+chumsky = { workspace = true }
 clap = { version = "4", features = ["derive"] }
-miette = { version = "7", features = ["fancy"] }
+miette = { workspace = true }
 thiserror = "2"
 tokio = { version = "1", features = ["full"] }
 reqwest = { version = "0.12", features = ["json"] }
