+++ crates/keel-syntax/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "keel-syntax"
+version = "0.2.1"
+edition = "2024"
+license = "MIT"
+description = "Keel front-end: lexer, AST, parser, formatter, and linter."
+publish = false
+
+[dependencies]
+logos = { workspace = true }
+chumsky = { workspace = true }
+miette = { workspace = true }
